Ignore comet request transactions in NewRelic.
Comets in Lift are long polls. As a result they tend to genuinely screw
with the ability for NewRelic to accurately count the response time of
the application. To combat this, we're going to ignore them for now.
Ideally we'll eventually have a way to *accurately* count the time spent
in an AJAX or Comet handler.